* changes:
  Loongson-3: LS3000/4000VZ: Extend guestid_cache and vpid_cache to unsigned long.
  Loongson-3: LS4000VZ: FTLB start at 64, no from (VTLBSIZE - 2), flush DIAG operation should not be interrupted.

+++ arch/mips/include/asm/cpu-info.h
@@ -114,9 +114,9 @@ struct cpuinfo_mips {
 	struct guest_info	guest;
 	unsigned int		gtoffset_mask;
 	unsigned int		guestid_mask;
-	unsigned int		guestid_cache;
+	unsigned long		guestid_cache;
 	/*3A3000 feature*/
-	unsigned int		vpid_cache;
+	unsigned long		vpid_cache;
 } __attribute__((aligned(SMP_CACHE_BYTES)));
 
 extern struct cpuinfo_mips cpu_data[];
